Fix a potential UB that likely causes issues on IBM Power9 architecture.

Passes STC
Elo   | -0.16 +- 2.76 (95%)
SPRT  | 10.0+0.10s Threads=1 Hash=16MB
LLR   | 3.06 (-2.94, 2.94) [-5.00, 0.00]
Games | N: 32374 W: 11970 L: 11985 D: 8419
Penta | [1491, 2674, 7821, 2761, 1440]
https://openbench.kaichung.dev/test/46/

+++ src/movegen/attacks.cpp
@@ -321,7 +321,7 @@ void populateRookAttacks() {
 }
 
 void populateBetweenSquares() {
-    uint8_t index;
+    int16_t index;
 
     for (uint8_t from = 0; from < 64; from++) {
         for (uint8_t to = from + 1; to < 64; to++) {
@@ -362,7 +362,7 @@ void populateBetweenSquares() {
 
 void populateLineSquares() {
     int8_t step;
-    uint8_t index;
+    int16_t index;
 
     for (uint8_t from = 0; from < 64; from++) {
         for (uint8_t to = from + 1; to < 64; to++) {
